update get ordered data to accept a key
This commits updates the get ordered data method
to allow it to accept a hash of the data to apply
it's rules against.

This is meant to handle a case where we might
want to perform one single lookup of all host
data in oder to reduce the amount of calls going
out to consul.