WIP: controlling upgrades
This commit is more psuedo-code than code.

Adds a method to control_updrades:

1 triggers an update (when required)
2 ensures that all nodes default to not-upgrading
3 passes a list of instructions to a method that
knows how to parse them and return a list of
actions to perform
4 passes that list of actions to a method that
performs them

NOTE: #3 is completely not implemented, I wrote
down some ideas about what it looks like, but I
still need to actually implement it and invent a way
to reasonably test it.