add key_status_off_role code
Currently, get_status just returns list of
hosts for each state (pending, upgraded, upgrading).

Code to perform updates needs this data in a different
format, namely keyed off role.

This commit adds a method to key status off role
so that it is easier for the upgrade code to
consume the status data.

it's possible that this means that the status data
is in the wrong format.